requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pyjwt==2.8.0
werkzeug==3.0.1